# backend/blog-adapters/web-rest/src/main/resources/static/openapi/blog-api.yml
openapi: 3.1.0
info:
  title: Blog API
  description: API for managing blog content (categories, tags, posts, author profiles).
  version: 0.0.1
servers:
  - url: http://localhost:8080
    description: Local server
...
paths:
...
  /api/tags:
    get:
      operationId: getTags
      tags: [ Tags ]
      summary: List tags (public - active tags only)
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Page index (0-based)
        - in: query
          name: size
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Page size
      responses:
        '200':
          description: List of tags
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageTagView'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
...
  /api/admin/tags:
    get:
      operationId: adminGetTags
      tags: [ Tags ]
      summary: List tags (admin only)
      security:
        - basicAuth: [ ]
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Page index (0-based)
        - in: query
          name: size
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Page size
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/TagStatus'
          required: false
          description: Filter by status (ACTIVE or ARCHIVED)
      responses:
        '200':
          description: List of tags
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageTagView'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    post:
      operationId: adminCreateTag
      tags: [ Tags ]
      summary: Create a tag (admin only)
      security:
        - basicAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTagRequest'
      responses:
        '201':
          description: Tag created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTagResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/admin/tags/{id}:
    get:
      operationId: adminGetTag
      tags: [ Tags ]
      summary: List tags (admin only)
      security:
        - basicAuth: []
      parameters:
        - in: path
          name: id
          description: Tag ID
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Tag
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagView'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    put:
      operationId: adminUpdateTag
      tags: [ Tags ]
      summary: Update a tag (admin only)
      security:
        - basicAuth: [ ]
      parameters:
        - in: path
          name: id
          required: true
          description: Tag ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTagRequest'
      responses:
        '204':
          description: Updated
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'

    delete:
      operationId: adminArchiveTag
      tags: [ Tags ]
      summary: Archive a tag (admin only)
      description: Marks the tag as ARCHIVED.
      security:
        - basicAuth: [ ]
      parameters:
        - in: path
          name: id
          required: true
          description: Tag ID
          schema:
            type: string
      responses:
        '204':
          description: Archived
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
...
components:
  securitySchemes:
    AdminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Must resolve to a user with ADMIN role
    basicAuth:
      type: http
      scheme: basic
      description: Basic authentication
  schemas:
...
    TagStatus:
      type: string
      enum: [ACTIVE, ARCHIVED]
    TagView:
      type: object
      required: [id, name, slug, status, createdAt]
      properties:
        id:
          type: string
          example: "ec02e0a6-5528-4661-b863-81171c3a98ff"
        name:
          type: string
          example: "DDD"
        slug:
          type: string
          example: "ddd"
        status:
          $ref: '#/components/schemas/TagStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type:
            - string
            - "null"
          format: date-time
        archivedAt:
          type:
            - string
            - "null"
          format: date-time
    PageTagView:
      type: object
      required: [ content, page, size, totalElements, totalPages ]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/TagView'
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 20
        totalElements:
          type: integer
          format: int64
          example: 42
        totalPages:
          type: integer
          example: 3
    CreateTagRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: "DDD"
    CreateTagResponse:
      type: object
      required: [id]
      properties:
        id:
          type: string
          example: "ca21ba83-1292-453a-b2fa-d7151df21dea"
    UpdateTagRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: "Distributed Systems"


    # ------------------------
    # POST
    # ------------------------
...
    ApiErrorResponse:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer, example: 400 }
        error: { type: string, example: "BAD_REQUEST" }
        message: { type: string, example: "Validation failed for request body." }
        path: { type: string, example: "/api/admin/path" }
        traceId: { type: string, example: "dbd2ebb9-f737-40e3-9cb4-ce1ec2b229f5" }
        details:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }